+++ home.nix
@@ -337,6 +337,7 @@ in {
 
   programs.direnv.enable = true;
   programs.direnv.nix-direnv.enable = true;
+  programs.direnv.nix-direnv.enableFlakes = true;
 
   programs.neovim = {
     enable = true;