+++ modules/samba.nix
@@ -1,5 +1,5 @@
 {
-  nixosModule = { ... }: {
+  nixosModule = { config, ... }: {
     services.samba.enable = true;
     services.samba.shares.phi = {
       path = "/";
@@ -10,5 +10,8 @@
       "force user" = "enzime";
       "force group" = "users";
     };
+
+    networking.firewall.allowedTCPPorts = (assert (!builtins.hasAttr "openFirewall" config.services.samba); [ 139 445 ]);
+    networking.firewall.allowedUDPPorts = (assert (!builtins.hasAttr "openFirewall" config.services.samba); [ 137 138 ]);
   };
 }